contacts/pyproject.toml
[tool.poetry]
name = "contacts"
version = "0.1.0"
description = ""
authors = ["ArthurMykhailiuk <artur.myhajlyuk@volo.global>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.37"
alembic = "^1.14.1"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
psycopg2-binary = "^2.9.10"
pydentic = "^0.0.1.dev3"
asyncpg = "^0.30.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
gravatar = "^0.1"
libgravatar = "^1.0.4"
python-multipart = "^0.0.20"
pydantic-settings = "^2.7.1"
fastapi-mail = "^1.4.2"
slowapi = "^0.1.9"
cloudinary = "^1.42.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
pytest-cov = "^6.0.0"
httpx = "^0.28.1"
aiosqlite = "^0.20.0"
redis = "^5.2.1"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"

[tool.pytest.ini_options]
pythonpath = ["."]
filterwarnings = "ignore::DeprecationWarning"
asyncio_default_fixture_loop_scope = "function"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
